+++ assets/javascripts/kitten/karl/organisms/header.js
@@ -46,7 +46,7 @@ const platformSwitchProps = {
   positionedWithBorder: true,
 
   buttonId: 'k-PlatformSwitch',
-  buttonClassName: 'k-ButtonIcon k-ButtonIcon--default k-ButtonIcon--dropdown ' +
+  buttonClassName: 'k-ButtonIcon k-ButtonIcon--default k-ButtonIcon--horizontal--arrow ' +
                    'k-Dropdown__button--rotationOnExpanded',
   buttonContentOnExpanded: unloggedHeaderPlatformSwitchButton,
   buttonContentOnCollapsed: unloggedHeaderPlatformSwitchButton,