+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Features:
 - Add prop `closeEvents` to close the dropdowns.
 - Fix button icons alignment (for Firefox)
 - Add new class `k-HorizontalNav__element`.
+- Add new class `k-HorizontalNav__element--fixedSize`.
 - Add new class `k-Header__item--fixedSize`.
 
 Fixes: