+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+Fixes:
+- Fix on Header.
+
 ## [v1.0.0] - 2016-11-16
 
 Features: