+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 Fixes:
 - Popover bottom navigation fixed.
+- Fix on Header.
 
 ## [v1.0.0] - 2016-11-16