+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 Features:
-- Add `onToggle` to Dropdowns
+- Add `onToggle` to Dropdowns.
 
 ## [v1.0.1] - 2016-11-16