+++ CHANGELOG.md
@@ -9,8 +9,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 Fixes:
 - Popover bottom navigation fixed.
 - Fix on Header.
-- Change class k-ButtonIcon--dropdown to k-ButtonIcon--horizontal--arrow.
-- Add class k-ButtonIcon--vertical--arrow.
+- Change class k-ButtonIcon--dropdown to k-ButtonIcon--horizontalArrow.
+- Add class k-ButtonIcon--verticalArrow.
 
 ## [v1.0.0] - 2016-11-16